run/animation/animator/requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-cloud-storage==2.9.0
gunicorn==21.2.0
langchain-google-genai==1.0.2
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
grpcio==1.59.0